+++ CSBidReceiptWindowController.m
@@ -69,7 +69,7 @@
 
 - (CSBidReceipt *)selectedBidReceipt {
 	CSBidReceipt *selectedBidReceipt = nil;
-	unsigned selectedBidReceiptIndex = [[self selectedBidReceiptIndexes] firstIndex];
+	NSUInteger selectedBidReceiptIndex = [[self selectedBidReceiptIndexes] firstIndex];
 	if (NSNotFound != selectedBidReceiptIndex && selectedBidReceiptIndex < [[self bidReceipts] count]) {
 		selectedBidReceipt = [[self bidReceipts] objectAtIndex:selectedBidReceiptIndex];
 	}
